requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
docx2txt==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
narwhals==1.35.0
nltk==3.9.1
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
phonenumbers==9.0.3
pillow==11.2.1
preshed==3.0.9
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resume-parser==0.8.4
rich==14.0.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stemming==1.0.1
streamlit==1.44.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tika==3.1.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wrapt==1.17.2
